# docs/api-docs.json
{"openapi":"3.0.0","paths":{"/v1/lead":{"post":{"operationId":"LeadController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateLeadDto"}}}},"responses":{"201":{"description":"The record has been successfully created.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Lead"}}}},"400":{"description":"Bad Request.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BadRequestEntity"}}}},"409":{"description":"Conflict.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConflictErrorEntity"}}}},"500":{"description":"Internal Server Error.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InternalServerErrorEntity"}}}}}},"get":{"operationId":"LeadController_findAll","parameters":[],"responses":{"200":{"description":"The records has been successfully retrieved.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Lead"}}}}}}}},"/v1/lead/{id}":{"get":{"operationId":"LeadController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The record has been successfully retrieved.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Lead"}}}},"403":{"description":"Forbidden.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForbiddenErrorEntity"}}}}}},"patch":{"operationId":"LeadController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateLeadDto"}}}},"responses":{"200":{"description":"The record has been successfully updated.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Lead"}}}}}},"delete":{"operationId":"LeadController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":"The record has been successfully deleted."}}}}},"info":{"title":"Nombre de la API","description":"Descripcion de la API","version":"Version de la API","contact":{"name":"Nombre del lider tecnico","url":"","email":"contacto@email.com"}},"tags":[],"servers":[{"url":"http://localhost:3000/api"},{"url":"prod url here"}],"components":{"schemas":{"CreateLeadDto":{"type":"object","properties":{"id":{"type":"string"},"rut":{"type":"string"},"dv":{"type":"string"},"primerNombre":{"type":"string"},"segundoNombre":{"type":"string"},"primerApellido":{"type":"string"},"segundoApellido":{"type":"string"},"telMovil":{"type":"string"},"canalRegistro":{"type":"string"},"sucursal":{"type":"string"},"clasificacion":{"type":"string","enum":["Frio","Tibio","Caliente"],"default":"Frio"},"estado":{"type":"string","enum":["SinGestion","Contactado","NoContactado","NoInteresado","Interesado"],"default":"SinGestion"},"tipoLead":{"type":"string","enum":["Empresa","Pensionado","Trabajador","Independiente"],"default":"Trabajador"},"direccion":{"type":"string"},"ciudad":{"type":"string"},"comuna":{"type":"string"},"region":{"type":"string"}},"required":["id","rut","dv","primerNombre","segundoNombre","primerApellido","segundoApellido","telMovil","canalRegistro","sucursal","clasificacion","estado","tipoLead","direccion","ciudad","comuna","region"]},"Lead":{"type":"object","properties":{"id":{"type":"integer","format":"int64","default":1},"rut":{"type":"string","default":"12345678"},"dv":{"type":"string","default":"9"},"primerNombre":{"type":"string","default":"Jhon"},"segundoNombre":{"type":"string","default":"Doe"},"primerApellido":{"type":"string","default":"Doe"},"segundoApellido":{"type":"string","default":"Doe"},"email":{"type":"string","default":"jhon.doe@email.cl"},"telMovil":{"type":"string","default":"+56912345678"},"canalRegistro":{"type":"string","default":"Canal de registro"},"sucursal":{"type":"string","default":"Sucursal"},"clasificacion":{"type":"string","enum":["Frio","Tibio","Caliente"],"default":"Frio"},"estado":{"type":"string","enum":["SinGestion","Contactado","NoContactado","NoInteresado","Interesado"],"default":"SinGestion"},"tipoLead":{"type":"string","enum":["SinGestion","Contactado","NoContactado","NoInteresado","Interesado"],"default":"SinGestion"},"fchRegistro":{"format":"date-time","type":"string","default":"2023-09-06T21:33:57.479Z"},"prioridad":{"type":"string","default":"priority"},"ejecutivoAsignado":{"type":"string","default":"ejecutivo"},"direccion":{"type":"string","default":"direccion"},"ciudad":{"type":"string","default":"ciudad"},"comuna":{"type":"string","default":"comuna"},"region":{"type":"string","default":"region"}},"required":["id","rut","dv","primerNombre","segundoNombre","primerApellido","segundoApellido","email","telMovil","canalRegistro","sucursal","clasificacion","estado","tipoLead","fchRegistro","prioridad","ejecutivoAsignado","direccion","ciudad","comuna","region"]},"BadRequestEntity":{"type":"object","properties":{"message":{"type":"string","description":"Error message","example":"Bad Request"},"code":{"type":"number","description":"Error code","example":400}},"required":["message","code"]},"ConflictErrorEntity":{"type":"object","properties":{"message":{"type":"string","description":"Error message","example":"Conflict Error"},"code":{"type":"number","description":"Error code","example":409}},"required":["message","code"]},"InternalServerErrorEntity":{"type":"object","properties":{"message":{"type":"string","description":"Error message","example":"Internal Server Error"},"code":{"type":"number","description":"Error code","example":500}},"required":["message","code"]},"ForbiddenErrorEntity":{"type":"object","properties":{"message":{"type":"string","description":"Error message","example":"Forbidden Error"},"code":{"type":"number","description":"Error code","example":403}},"required":["message","code"]},"UpdateLeadDto":{"type":"object","properties":{"id":{"type":"string"},"rut":{"type":"string"},"dv":{"type":"string"},"primerNombre":{"type":"string"},"segundoNombre":{"type":"string"},"primerApellido":{"type":"string"},"segundoApellido":{"type":"string"},"telMovil":{"type":"string"},"canalRegistro":{"type":"string"},"sucursal":{"type":"string"},"clasificacion":{"type":"string","enum":["Frio","Tibio","Caliente"],"default":"Frio"},"estado":{"type":"string","enum":["SinGestion","Contactado","NoContactado","NoInteresado","Interesado"],"default":"SinGestion"},"tipoLead":{"type":"string","enum":["Empresa","Pensionado","Trabajador","Independiente"],"default":"Trabajador"},"direccion":{"type":"string"},"ciudad":{"type":"string"},"comuna":{"type":"string"},"region":{"type":"string"}}}}}}